Deprecate DashboardNavbarItem in favor of DashboardModule
Rename new DashboardTabModule to DashboardModule
