+++ misk-admin/src/main/kotlin/misk/web/dashboard/DashboardNavbarItem.kt
@@ -7,12 +7,15 @@ import misk.web.dashboard.ValidWebEntry.Companion.slugify
  * Misk-Web's default navbar handles responsive layout of these items so overflow
  *   navbar items are included in the drop down menu
  */
+@Deprecated("DashboardNavbarStatus will not be supported in Misk Dashboards v2 and will be deleted shortly. Use DashboardModule.createMenuLink() instead.")
+
 data class DashboardNavbarItem(
   val dashboard_slug: String,
   val item: String,
   val order: Int
 ) : ValidWebEntry(valid_slug = dashboard_slug)
 
+@Deprecated("DashboardNavbarStatus will not be supported in Misk Dashboards v2 and will be deleted shortly. Use DashboardModule.createMenuLink() instead.")
 inline fun <reified DA : Annotation> DashboardNavbarItem(
   item: String,
   order: Int
